+++ CHANGELOG.rst
@@ -1,4 +1,14 @@
 Changelog for Nautilus
+===================================
+
+0.1.9 (2023-02-17)
+-------------------
+
+Added:
+^^^^^^
+- Store TIFF files under subdirectory in the format prefix__YYYY_M_D_HMS
+- Add settings.txt file to each acquisition with led intensity, stage positions, duration, and frame rate
+
 ======================
 
 0.1.8 (2023-02-17)